+++ src/main/java/td_game/model/enemy/IDamageable.java
@@ -1,6 +1,6 @@
 package td_game.model.enemy;
 
 public interface IDamageable {
-    void takeDamage(double amount);
+    void takeDamage(int amount);
     boolean isAlive();
 }
